implement: トップページを検索画面に変更。。
