--- youkai-front/src/pages/foundpassword.vue
@@ -1,18 +1,0 @@
-<template>
-  <v-layout>
-    <FoundPasswordSearch />
-  </v-layout>
-</template>
-
-<script lang="ts">
-import { Component, Vue } from 'vue-property-decorator'
-import FoundPasswordSearch from '@/components/foundpasswordsearch/FoundPasswordSearch.vue'
-
-@Component({
-  components: {
-    FoundPasswordSearch,
-  },
-})
-export default class FoundPassword extends Vue {
-}
-</script>
